refactor: 전략, 지표들의 폴더 구조 수정
