+++ src/services/analysis/indicators/bollinger/bollinger.types.ts
