+++ src/services/analysis/indicators/bollinger/bollinger.utils.ts
