+++ src/services/analysis/indicators/ma/ma.queries.ts
