+++ src/services/analysis/indicators/ma/ma.types.ts
