+++ src/services/analysis/indicators/macd/macd.queries.ts
