+++ src/services/analysis/indicators/macd/macd.types.ts
