+++ src/services/analysis/indicators/rsi/rsi.types.ts
