+++ src/services/analysis/indicators/stochastic/stochastic.queries.ts
