+++ src/services/analysis/indicators/stochastic/stochastic.types.ts
