+++ src/services/analysis/indicators/stochastic/stochastic.utils.ts
