+++ src/services/analysis/indicators/volume/volume.queries.ts
