--- src/strategy/iStrategy.ts
@@ -1,23 +1,0 @@
-import type { PoolClient } from "pg";
-
-/**
- * Strategy Interface
- * Defines the structure for implementing trading strategies
- *
- * 전략 인터페이스
- * 거래 전략을 구현하기 위한 구조를 정의
- */
-export interface iStrategy {
-	client: PoolClient;
-	uuid: string;
-	symbol: string;
-	weight: number;
-
-	execute(): Promise<number>;
-}
-
-export enum Signal {
-	BUY = "BUY",
-	SELL = "SELL",
-	HOLD = "HOLD",
-}
